+++ ootrando_overworldmap_hamsda/changelog.md
@@ -5,6 +5,7 @@
 - fixed closed DoT logic for `DMT Freestanding PoH`
 - changed `Randomize Spawns` setting to have separate options for child/adult/both
 - added `Shuffle Frog Song Rupees` setting
+- added `Blue Fire Arrows` setting
 
 ## 3.7.3.0